+++ src/modules/session/session.service.ts
@@ -1,17 +1,21 @@
-import { Injectable } from '@nestjs/common';
-import { databaseHandler } from 'src/database/database.handler';
+import { Injectable, BadRequestException, NotFoundException } from '@nestjs/common';
 import { DatabaseService } from 'src/database/database.service';
+import { handlers } from 'src/handlers/handlers';
 
 @Injectable()
 export class SessionService {
   constructor(private readonly databaseService: DatabaseService) {}
 
   async getAll(userId: string) {
-    return await this.getAllInDatabase(userId);
+    const res = await this.getAllInDatabase(userId);
+    return res;
   }
 
   async delete(id: string, userId: string) {
-    return await this.deleteInDatabase(id, userId);
+    const resFindSession = await this.findSessionInDb(id, userId);
+    if (!resFindSession) throw new NotFoundException('Session is not found');
+    await this.deleteInDatabase(id, userId);
+    return { message: 'Session is delete' };
   }
 
   // ----------------------------------------------------------------------
@@ -23,15 +27,26 @@ export class SessionService {
   // ----------------------------------------------------------------------
 
   private async getAllInDatabase(userId: string) {
-    return await databaseHandler.errors(
+    if (!userId) throw new BadRequestException();
+    return await handlers.dbError(
       this.databaseService.session.findMany({
         where: { userId },
       }),
     );
   }
 
+  private async findSessionInDb(id: string, userId: string) {
+    if (!id || !userId) throw new BadRequestException();
+    return await handlers.dbError(
+      this.databaseService.session.findFirst({
+        where: { id, userId },
+      }),
+    );
+  }
+
   private async deleteInDatabase(id: string, userId: string) {
-    return await databaseHandler.errors(
+    if (!id || !userId) throw new BadRequestException();
+    return await handlers.dbError(
       this.databaseService.session.delete({
         where: { id, userId },
       }),
