+++ src/modules/user/user.service.ts
@@ -1,7 +1,7 @@
 import { Injectable, BadRequestException } from '@nestjs/common';
 import { DatabaseService } from '../../database/database.service';
-import { databaseHandler } from '../../database/database.handler';
 import { password } from '../../utils/password';
+import { handlers } from 'src/handlers/handlers';
 
 interface IDeleteUser {
   userId: string;
@@ -21,31 +21,36 @@ export class UserService {
   /* ----------------  GET  ---------------- */
 
   async myAccount(userId: string) {
-    const userInDb = await this.findUserById(userId);
-    return userInDb;
+    const { id, name, roleId } = await this.findUserById(userId);
+    return { id, name, roleId };
   }
 
   /* ----------------  PUT  ---------------- */
 
   // edit name
   async editName(userId: string, name: string) {
-    const editResult = await this.editNameInDb({ userId, newName: name });
-    return editResult;
+    const userInDb = await this.findUserById(userId)
+    if (userInDb.name === name) throw new BadRequestException('Error edit user name')
+    await this.editNameInDb({ userId, newName: name });
+    return { message: 'Name is edit' };
   }
 
   // edit password
   async editPassword(data: IEditPassword) {
     const userInDb = await this.findUserById(data.userId);
 
+    // check user password
     const checkPassword = await password.verify(data.oldPassword, userInDb.password);
+    const checkPasswordDuplicate = await password.verify(data.newPassword, userInDb.password);
     if (!checkPassword) throw new BadRequestException('The password is not correct');
+    if (checkPasswordDuplicate) throw new BadRequestException('The password is already set');
 
     const newPassword = await password.generateHash(data.newPassword);
-    const resEditPassword = await this.editPasswordInDb({ newPassword, userId: data.userId });
 
+    await this.editPasswordInDb({ newPassword, userId: data.userId });
     await this.deleteSessionsInDb(data.userId);
 
-    return resEditPassword;
+    return { message: 'Password is edit' };
   }
 
   /* ----------------  DELETE  ---------------- */
@@ -62,6 +67,8 @@ export class UserService {
     await this.editIssuesInDb(data.userId);
     await this.editIssuesCommentInDb(data.userId);
     await this.deleteUserInDb(data.userId);
+
+    return { message: 'User is delete' };
   }
 
   // ----------------------------------------------------------------------
@@ -73,7 +80,8 @@ export class UserService {
   // ----------------------------------------------------------------------
 
   private async findUserById(userId: string) {
-    return databaseHandler.errors(
+    if (!userId) throw new BadRequestException();
+    return handlers.dbError(
       this.databaseService.user.findFirst({
         where: { id: userId },
       }),
@@ -81,7 +89,8 @@ export class UserService {
   }
 
   private async editNameInDb({ userId, newName }: { userId: string; newName: string }) {
-    return databaseHandler.errors(
+    if (!userId || !newName) throw new BadRequestException();
+    return handlers.dbError(
       this.databaseService.user.update({
         where: { id: userId },
         data: { name: newName },
@@ -90,7 +99,8 @@ export class UserService {
   }
 
   private async editPasswordInDb({ userId, newPassword }: { userId: string; newPassword: string }) {
-    return databaseHandler.errors(
+    if (!userId || !newPassword) throw new BadRequestException();
+    return handlers.dbError(
       this.databaseService.user.update({
         where: { id: userId },
         data: { password: newPassword },
@@ -99,7 +109,8 @@ export class UserService {
   }
 
   private async deleteSessionsInDb(userId: string) {
-    return databaseHandler.errors(
+    if (!userId) throw new BadRequestException();
+    return handlers.dbError(
       this.databaseService.session.deleteMany({
         where: { userId },
       }),
@@ -107,7 +118,8 @@ export class UserService {
   }
 
   private async deleteUserInDb(userId: string) {
-    return databaseHandler.errors(
+    if (!userId) throw new BadRequestException();
+    return handlers.dbError(
       this.databaseService.user.delete({
         where: { id: userId },
       }),
@@ -115,7 +127,8 @@ export class UserService {
   }
 
   private async deleteUserPassCollectionInDb(userId: string) {
-    return databaseHandler.errors(
+    if (!userId) throw new BadRequestException();
+    return handlers.dbError(
       this.databaseService.passCollection.deleteMany({
         where: { userId },
       }),
@@ -123,7 +136,8 @@ export class UserService {
   }
 
   private async deleteUserConfigsInDb(userId: string) {
-    return databaseHandler.errors(
+    if (!userId) throw new BadRequestException();
+    return handlers.dbError(
       this.databaseService.config.deleteMany({
         where: { userId },
       }),
@@ -131,7 +145,8 @@ export class UserService {
   }
 
   private async editIssuesInDb(userId: string) {
-    return databaseHandler.errors(
+    if (!userId) throw new BadRequestException();
+    return handlers.dbError(
       this.databaseService.issue.updateMany({
         where: { userId },
         data: { userId: null },
@@ -140,7 +155,8 @@ export class UserService {
   }
 
   private async editIssuesCommentInDb(userId: string) {
-    return databaseHandler.errors(
+    if (!userId) throw new BadRequestException();
+    return handlers.dbError(
       this.databaseService.issueComment.updateMany({
         where: { userId },
         data: { userId: null },
